filesizes >2gb fixed
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@2073 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/gui/skins/simple/DownloadInfo.java
@@ -198,9 +198,9 @@ public class DownloadInfo extends JFrame {
                 JProgressBar p;
                 Chunk next = it.next();
                 addEntry(JDLocale.L("download.chunks.connection", "Verbindung") + " " + i, p = new JProgressBar(0, 100));
-                p.setValue(next.getBytesLoaded() * 100 / Math.max(1, next.getChunkSize()));
+                p.setValue((int)(next.getBytesLoaded() * 100 / Math.max(1, next.getChunkSize())));
                 p.setStringPainted(true);
-                p.setString(JDUtilities.formatKbReadable((int) next.getBytesPerSecond() / 1024) + "/s " + JDUtilities.getPercent(next.getBytesLoaded(), next.getChunkSize()));
+                p.setString(JDUtilities.formatKbReadable(next.getBytesPerSecond() / 1024) + "/s " + JDUtilities.getPercent(next.getBytesLoaded(), next.getChunkSize()));
             }
 
         }
